feat: add monitors to pages

+++ packages/api/src/router/page.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { eq } from "@openstatus/db";
+import { and, eq, inArray } from "@openstatus/db";
 import {
   insertPageSchema,
   insertPageSchemaWithMonitors,
@@ -10,6 +10,8 @@ import {
   selectIncidentSchema,
   selectMonitorSchema,
   selectPageSchema,
+  user,
+  usersToWorkspaces,
 } from "@openstatus/db/src/schema";
 
 import { createTRPCRouter, protectedProcedure, publicProcedure } from "../trpc";
@@ -39,8 +41,24 @@ export const pageRouter = createTRPCRouter({
   getPageById: protectedProcedure
     .input(z.object({ id: z.number() }))
     .query(async (opts) => {
+      const currentUser = await opts.ctx.db
+        .select()
+        .from(user)
+        .where(eq(user.tenantId, opts.ctx.auth.userId))
+        .get();
+
+      const result = await opts.ctx.db
+        .select()
+        .from(usersToWorkspaces)
+        .where(eq(usersToWorkspaces.userId, Number(currentUser.id)))
+        .all();
+      const workspaceIds = result.map((workspace) => workspace.workspaceId);
+
       return await opts.ctx.db.query.page.findFirst({
-        where: eq(page.id, opts.input.id),
+        where: and(
+          eq(page.id, opts.input.id),
+          inArray(page.workspaceId, workspaceIds),
+        ),
         with: { monitorsToPages: { with: { monitor: true } }, incidents: true },
       });
     }),
@@ -69,15 +87,57 @@ export const pageRouter = createTRPCRouter({
   deletePage: protectedProcedure
     .input(z.object({ pageId: z.number() }))
     .mutation(async (opts) => {
+      const currentUser = await opts.ctx.db
+        .select()
+        .from(user)
+        .where(eq(user.tenantId, opts.ctx.auth.userId))
+        .get();
+
+      const result = await opts.ctx.db
+        .select()
+        .from(usersToWorkspaces)
+        .where(eq(usersToWorkspaces.userId, Number(currentUser.id)))
+        .all();
+      const workspaceIds = result.map((workspace) => workspace.workspaceId);
+
+      const pageToDelete = await opts.ctx.db
+        .select()
+        .from(page)
+        .where(
+          and(
+            eq(page.id, opts.input.pageId),
+            inArray(page.workspaceId, workspaceIds),
+          ),
+        )
+        .get();
+      if (!pageToDelete) return;
       await opts.ctx.db.delete(page).where(eq(page.id, opts.input.pageId));
     }),
   getPageByWorkspace: protectedProcedure
     .input(z.object({ workspaceId: z.number() }))
     .query(async (opts) => {
+      const currentUser = await opts.ctx.db
+        .select()
+        .from(user)
+        .where(eq(user.tenantId, opts.ctx.auth.userId))
+        .get();
+
+      const result = await opts.ctx.db
+        .select()
+        .from(usersToWorkspaces)
+        .where(eq(usersToWorkspaces.userId, Number(currentUser.id)))
+        .all();
+      const workspaceIds = result.map((workspace) => workspace.workspaceId);
+
       return opts.ctx.db
         .select()
         .from(page)
-        .where(eq(page.workspaceId, opts.input.workspaceId))
+        .where(
+          and(
+            eq(page.workspaceId, opts.input.workspaceId),
+            inArray(page.workspaceId, workspaceIds),
+          ),
+        )
         .all();
     }),
 
@@ -85,15 +145,34 @@ export const pageRouter = createTRPCRouter({
   getPageBySlug: publicProcedure
     .input(z.object({ slug: z.string() }))
     .query(async (opts) => {
-      const result = opts.ctx.db.query.page.findFirst({
+      const result = await opts.ctx.db.query.page.findFirst({
         where: eq(page.slug, opts.input.slug),
-        with: { monitorsToPages: { with: { monitor: true } }, incidents: true },
+        with: { incidents: true },
       });
+
+      if (!result) {
+        return;
+      }
+
+      // FIXME: There is probably a better way to do this
+      const monitorsToPagesResult = await opts.ctx.db
+        .select()
+        .from(monitorsToPages)
+        .where(eq(monitorsToPages.pageId, result.id))
+        .all();
+      const monitorsId = monitorsToPagesResult.map(
+        (monitor) => monitor.monitorId,
+      );
+      const monitors = await opts.ctx.db
+        .select()
+        .from(monitor)
+        .where(inArray(monitor.id, monitorsId))
+        .all();
       const selectPageSchemaWithRelation = selectPageSchema.extend({
         monitors: z.array(selectMonitorSchema),
         incidents: z.array(selectIncidentSchema),
       });
 
-      return selectPageSchemaWithRelation.parse(result);
+      return selectPageSchemaWithRelation.parse({ ...result, monitors });
     }),
 });
